Extract ProjectCommandSet into its own file
